don't try hdr with firefox.
We make an call to getContextAttributes(), which will lead to an error
that we then catch when using firefox (until HDR functions in firefox
are implemented). Only using getContext with colorSpace argument will
not make it fail, which is why in the current version, firefox always
seems to provide an HDR canvas. This is not the case, so we make it
fail.